Broke out fixtures to conftest and constants to a separate file
